+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/brevdev/vault-utils/compare/v0.1.0...HEAD)
 
+### Added 
+
+- Added proper exit codes if error instead of panics.
+- Added more robust validation messages
+
 ### Changed
 
 - Replaced template readme with vault-utils readme